Break our header icon into its own file
