+++ src/Header.module.css
@@ -43,252 +43,6 @@ body {
   --header-height: 50px;
 }
 
-.headerIcon {
-  position: fixed;
-	display: flex;
-  top: 0;
-  left: 0;
-  user-select: none;
-  cursor: pointer;
-  z-index: 2;
-}
-
-.headerIcon .logo {
-  width: var(--header-height);
-	height: var(--header-height);
-  margin-left: 10px;
-  user-select: none;
-}
-.headerIcon .logo img {
-  pointer-events: none;
-}
-.headerIcon .logo:hover img {
-  transform: scale(1.4);
-}
-
-.headerIcon .characterIcon {
-  display: flex;
-  margin-top: 10px;
-  margin-left: 10px;
-  transform: translate(0, 0);
-  transition: transform 0.5s cubic-bezier(0, 1, 0, 1);
-}
-.headerIcon .characterIcon .canvas {
-  --canvas-size: 80px;
-
-  position: relative;
-  width: var(--canvas-size);
-  height: var(--canvas-size);
-  background-color: #FFF;
-  border: 3px solid #000;
-  border-radius: 1000px;
-  z-index: 1;
-}
-.headerIcon .characterIcon .canvas:not(:first-child) {
-  display: none;
-}
-.headerIcon .characterIcon .meta {
-  display: flex;
-  flex-direction: column;
-  width: 150px;
-}
-.headerIcon .characterIcon .meta .text {
-  position: relative;
-  display: flex;
-  margin-left: -50px;
-  padding: 5px 10px;
-  padding-left: 50px;
-  font-family: WinchesterCaps;
-  font-size: 16px;
-  color: #FFF;
-  letter-spacing: 1px;
-  align-items: baseline;
-}
-.headerIcon .characterIcon .meta .text .name,
-.headerIcon .characterIcon .meta .text .level
-{
-  position: relative;
-}
-.headerIcon .characterIcon .meta .text .name {
-  margin-right: auto;
-}
-.headerIcon .characterIcon .meta .text .level {
-  font-size: 12px;
-}
-.headerIcon .characterIcon .meta .text .background {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-color: #000;
-  transform: skewX(-30deg);
-  overflow: hidden;
-}
-.headerIcon .characterIcon .meta .text .background::before {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-color: #42a5f5;
-  transform: translateX(-100%);
-  transition: transform 0.5s cubic-bezier(0, 1, 0, 1);
-  content: '';
-}
-.headerIcon .characterIcon .meta .stat {
-  display: flex;
-  position: relative;
-}
-.headerIcon .characterIcon .meta .stat .label {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  width: 52px;
-  margin-left: -30px;
-  padding: 0 5px;
-  padding-left: 30px;
-  background-color: #000;
-  color: #FFF;
-  font-family: 'PlazaRegular';
-  font-size: 14px;
-}
-.headerIcon .characterIcon .meta .stat .value {
-  position: absolute;
-  top: 0;
-  left: 22px;
-  /* font-family: 'PlazaRegular'; */
-  font-family: 'WinchesterCaps';
-  font-size: 13px;
-  letter-spacing: 1px;
-  color: #FFF;
-  text-shadow: 1px 1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, -1px -1px 0 #000;
-}
-.headerIcon .characterIcon .meta .stat .progress {
-  width: 100%;
-  height: 3px;
-  margin-top: auto;
-  margin-bottom: 3px;
-  appearance: none;
-}
-.headerIcon .characterIcon .meta .stat .progress::-webkit-progress-bar {
-  background: transparent;
-}
-.headerIcon .characterIcon .meta .stat.hp .progress::-webkit-progress-value {
-  background-image: linear-gradient(to right, #ad1457, #f48fb1);
-}
-.headerIcon .characterIcon .meta .stat.mp .progress::-webkit-progress-value {
-  background-image: linear-gradient(to right, #7cb342, #9ccc65);
-}
-.headerIcon .characterIcon .meta .stat .progress::-webkit-progress-value {
-  background-image: linear-gradient(to right, #1e88e5, #42a5f5);
-}
-
-.headerIcon .characterIcon .meta .stat .barImg {
-  position: absolute;
-  width: 100%;
-  left: 0;
-  bottom: 0;
-}
-
-.headerIcon .characterIcon .main {
-  display: flex;
-  z-index: 1;
-}
-.headerIcon .characterIcon .sub {
-  position: absolute;
-  display: flex;
-  top: 81px;
-  left: 0;
-}
-.headerIcon .characterIcon .limitBar {
-  position: relative;
-  height: 13px;
-  width: 150px;
-  margin-top: 7px;
-  /* margin-left: 20px; */
-  border: 3px solid;
-  background: linear-gradient(to bottom,rgb(46, 46, 46) 10%,rgb(116, 116, 116) 40%,rgb(116, 116, 116) 100%);
-}
-.headerIcon .characterIcon .limitBar .label {
-  position: absolute;
-  top: 0;
-  font-size: 10px;
-  /* line-height: 1; */
-  letter-spacing: 1px;
-  color: #FFF;
-  /* -webkit-text-stroke: 0.5px #000; */
-  text-shadow: 1px 1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, -1px -1px 0 #000;
-  text-transform: uppercase;
-}
-.headerIcon .characterIcon .limitBar .inner {
-  position: absolute;
-  /* margin: 3px 0px 0px 23px; */
-  /* height: 13px; */
-  top: 0;
-  bottom: 0;
-  left: 0;
-  width: 50px;
-  background: linear-gradient(to bottom,rgb(194, 130, 130) 20%,rgb(255, 255, 255) 40%,rgb(255, 170, 170) 60%, rgb(255, 170, 170) 100%);
-}
-.headerIcon .buttonWrap {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 80px;
-  color: #FFF;
-  font-size: 12px;
-  /* background: transparent; */
-  /* background: #000; */
-  /* border: 0; */
-  border-radius: 3px;
-}
-.headerIcon .buttonWrap .button {
-  padding: 2px;
-  background-color: #000;
-  /* border: 1px solid #fff; */
-  border-radius: 3px;
-  transition: background-color 0.5s cubic-bezier(0, 1, 0, 1);
-}
-.headerIcon:hover .characterIcon {
-  transform: translateX(10px);
-}
-/* .headerIcon:hover .main {
-  transform: scale(1.1, 1.1);
-} */
-.headerIcon:hover .buttonWrap .button {
-  background-color: #42a5f5;
-  /* color: #FFF; */
-}
-.headerIcon:hover .characterIcon .meta .text .background::before {
-  transform: translate(0, 0);
-}
-/* .characterBtn:hover .background, .characterWrapper.opened .background {
-  transform: rotateZ(180deg) translateX(50px);
-}
-.characterBtn .btnShortKey {
-  position: relative;
-  z-index: 2;
-}
-.characterBtn .btnShortKey {
-  color: #fff;
-  border: 1px solid #fff;
-  border-radius: 3px;
-  font-size: 13px;
-  padding: 2px;
-  margin-left: 4px;
-  z-index: 2;
-}
-.characterBtn:hover .btnShortKey, .characterWrapper.opened .btnShortKey {
-  border-color: #000;
-  color: #000;
-}
-.characterBtn:hover .btnText, .characterWrapper.opened .btnText {
-  position: relative;
-  color: #000;
-  z-index: 2;
-} */
-
 .header .tab {
   position: fixed;
   transition: transform 0.3s cubic-bezier(0, 1, 0, 1);
